requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
google-api-python-client
youtube-transcript-api
openai>=1.0.0
python-dotenv
pydantic==2.10.6


